third_party/WebKit/LayoutTests/http/tests/eventsource/eventsource-cors-basic-expected.txt

CONSOLE ERROR: EventSource cannot load http://localhost:8000/eventsource/resources/es-cors-basic.php?count=1. No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: EventSource cannot load http://localhost:8000/eventsource/resources/es-cors-basic.php?count=2. The 'Access-Control-Allow-Origin' whitelists only 'http://some.other.origin:80'. Origin 'http://127.0.0.1:8000' is not in the list, and is therefore not allowed access.
Test that basic EventSource cross-origin requests fail until they are allowed by the Access-Control-Allow-Origin header. Should print a series of PASS messages followed by DONE.

PASS: got error event and readyState is CLOSED
PASS: got error event and readyState is CLOSED
PASS: got cross-origin message event with data "DATA1" and lastEventId "77"
PASS: got cross-origin message event with data "DATA2"
DONE
